papers/3_moral_geometry/: fix hidden_dim derivation and MPS float16 cast in 5 scripts
Derive hidden_dim from NPZ probe files instead of non-existent
model.info.hidden_dim. Add .float() cast to mean-pooled activations
for MPS compatibility (returns float16, probes expect float32).

+++ papers/3_moral_geometry/scripts/dilemma_fragility.py
@@ -58,7 +58,7 @@ def collect_activations(model, texts: list[str], n_layers: int) -> dict[int, lis
         acts = model.get_activations(text, layers=list(range(n_layers)))
         for layer_idx in range(n_layers):
             h = acts[layer_idx]
-            pooled = h.mean(dim=1).squeeze(0)
+            pooled = h.mean(dim=1).squeeze(0).float()
             all_acts[layer_idx].append(pooled.cpu())
     return all_acts
